don't require that plugins are able to fulfill timestep requests immediately
notify plugins that displaying another timestep is desired and allow for some
time to handle that request

requestTimestep plugin method will be called when another timestep is to be
prepared. plugin should call commitTimestep as soon as it is ready. when all
plugins are ready, timestep will be applied and all plugin's setTimestep
method will be called.