testScenario.xosc with includes and array
put included files in separate directory

the included files are written to a file with _out.xml at the end of the
file name

Attention:
Only the file that is included from the main file directly is written to
the correct directory.
The other included files are written to the main directory instead of
the directory from where they are read.