Add ability to enter address for google map
Previously, the loadGoogleMap() method could only take coordinates as
parameters. Using Google's own geocoding API, the method now downloads
an XML document with the geocode of the entered address, and parses it
to find the coordinates of the entered address. Requiring the user to
enter coordinates when they want to load a map is cumbersome, this
change should make the process more natural.

In addition, previously the map would only load correctly if the view
was at the standard zoom. This is still the cause, but now the
loadGoogleMap() method zooms to the standard zoom before loading.