Shrink images for google map, add new save scheme
The images added to the scene by the loadGoogleMap method are now scaled
to be much smaller, without losing image quality. The images now adhere
to real-life scale much more closely.

Previously, the images were deleted after they were added. This caused
problems when loading an .xodr file later, as the images are not saved
in the .xodr file itself, only their filenames. The program now saves
the images in a schema that makes sense (in a separate
"OddlotMapsImages" folder), according to location, maptype, and size.