Fix OSM import, Add audio receiver control
In Oddlot, an error would occur when importing OSM files with nodes
greater than the maximum value for an unsigned int. Changed the process
to use a floating point instead, so the error should no longer occur.

In addition, when Google Maps images are added to the scene in oddlot,
they are now placed at the correct x,y coordinates in the scene, based
on the latitude and longitude used to pull the images.

Finally, hlrsDemo, which controls the 3x3 TV grid in the VIS lab, has
been updated to also support sending certain commands to the audio
receiver in the VIS cluster, via IR.