save per-model settings to .cover sidecar file
When a file is loaded by a stand-alone cover (not part of a
covise/vistle session), a config::Bridge is provided so that per-model
settings will be stored to and restored from a .toml file. This file is
stored in the same directory as the data set and has .cover appended to
its name.