re-add symbol export
When compiling with hidden visiblity, using operator<< on a VrmlNode
from outside the vrml library proper requires that this operator is
exported. So re-add the export, but this time before the friend
declaration, assuming that the compile problems on windows stem from the
fact that the first declaration (as a friend) did not specify all
decorators.