add .editorconfig
This config file, as specified at https://editorconfig.org/, helps many
editors and IDEs to automatically apply the correct settings, such as
encoding, EOL characters, indentation and final newlines, to be applied
for files in a project. I've configured it to match the existing style,
such that most editors will follow the same rules and produce no diff
when loading and re-saving a code file.